+++ src/mindlessgen/prog/config.py
@@ -6,6 +6,8 @@ from __future__ import annotations
 
 from pathlib import Path
 from abc import ABC, abstractmethod
+import warnings
+import multiprocessing as mp
 import toml
 
 from ..molecules import PSE_NUMBERS
@@ -780,6 +782,70 @@ class ConfigManager:
 
         if config_file:
             self.load_from_toml(config_file)
+
+    def check_config(self, verbosity: int = 1) -> None:
+        """
+        Checks ConfigClass for any incompatibilities that are imaginable
+        """
+
+        # lower number of the available cores and the configured parallelism
+        num_cores = min(mp.cpu_count(), self.general.parallel)
+        if self.general.parallel > mp.cpu_count() and verbosity > -1:
+            warnings.warn(
+                f"Number of cores requested ({self.general.parallel}) is greater "
+                + f"than the number of available cores ({mp.cpu_count()})."
+                + f"Using {num_cores} cores instead."
+            )
+
+        if num_cores > 1 and self.postprocess.debug and verbosity > -1:
+            # raise warning that debugging of postprocessing will disable parallelization
+            warnings.warn(
+                "Debug output might seem to be redundant due to the parallel processes "
+                + "with possibly similar errors in parallel mode. "
+                + "Don't be confused!"
+            )
+
+        if num_cores > 1 and verbosity > 0:
+            # raise warning that parallelization will disable verbosity
+            warnings.warn(
+                "Parallelization will disable verbosity during iterative search. "
+                + "Set '--verbosity 0' or '-P 1' to avoid this warning, or simply ignore it."
+            )
+
+        if self.refine.engine == "xtb":
+            # Check for f-block elements in forbidden elements
+            if self.generate.forbidden_elements:
+                if verbosity > 0:
+                    f_block_elements = set(range(56, 71)) | set(range(88, 103))
+                    lanthanides = set(range(56, 71))
+                    if not all(
+                        elem in self.generate.forbidden_elements for elem in lanthanides
+                    ) or any(
+                        elem in f_block_elements
+                        for elem in self.generate.element_composition
+                    ):
+                        warnings.warn(
+                            "f-block elements could be within the molecule. "
+                            + "xTB does not treat f electrons explicitly. "
+                            + "UHF is temporarily set to 0."
+                        )
+
+            # Check for super heavy elements in forbidden elements
+            super_heavy_elements = set(range(86, 102))
+            if self.generate.element_composition and any(
+                elem in super_heavy_elements
+                for elem in self.generate.element_composition
+            ):
+                if verbosity > 0:
+                    warnings.warn(
+                        "xTB does not treat super heavy elements. Atomic numbers are temporarily reduced by 32 to their lighter homologues and then replaced with the correct atom number."
+                    )
+                # Check if postprocessing is turned off
+                if not self.general.postprocess:
+                    if verbosity > 0:
+                        warnings.warn(
+                            "Postprocessing is turned off. The structure will not be relaxed."
+                        )
 
     def get_all_identifiers(self):
         """