+++ src/mindlessgen/qm/orca.py
@@ -165,6 +165,9 @@ class ORCA(QMMethod):
         orca_input = f"! {self.cfg.functional} {self.cfg.basis}\n"
         orca_input += f"! DEFGRID{self.cfg.gridsize}\n"
         orca_input += "! NoTRAH NoSOSCF SlowConv\n"
+        # "! AutoAux" keyword for super-heavy elements as def2/J ends at Rn
+        if any(atom >= 86 for atom in molecule.ati):
+            orca_input += "! AutoAux\n"
         if optimization:
             orca_input += "! OPT\n"
             if opt_cycles is not None: