+++ src/mindlessgen/qm/xtb.py
@@ -36,6 +36,11 @@ class XTB(QMMethod):
         """
         Optimize a molecule using xtb.
         """
+        super_heavy_elements = False
+        ati_original = molecule.ati.copy()
+        if np.any(molecule.ati > 85):
+            super_heavy_elements = True
+            molecule.ati[molecule.ati > 85] -= 32
         if np.any(np.isin(molecule.ati, get_lanthanides())):
             check_ligand_uhf(molecule.ati, molecule.charge)
             # Store the original UHF value and set uhf to 0
@@ -82,12 +87,21 @@ class XTB(QMMethod):
             if np.any(np.isin(molecule.ati, get_lanthanides())):
                 # Reset the UHF value to the original value before returning the optimized molecule.
                 optimized_molecule.uhf = uhf_original
+            if super_heavy_elements:
+                # Reset the atomic numbers to the original values before returning the optimized molecule.
+                optimized_molecule.ati = ati_original
+                optimized_molecule.atlist = molecule.atlist
             return optimized_molecule
 
     def singlepoint(self, molecule: Molecule, verbosity: int = 1) -> str:
         """
         Perform a single-point calculation using xtb.
         """
+        super_heavy_elements = False
+        ati_original = molecule.ati.copy()
+        if np.any(molecule.ati > 85):
+            super_heavy_elements = True
+            molecule.ati[molecule.ati > 85] -= 32
         if np.any(np.isin(molecule.ati, get_lanthanides())):
             check_ligand_uhf(molecule.ati, molecule.charge)
             # Store the original UHF value and set uhf to 0
@@ -128,6 +142,9 @@ class XTB(QMMethod):
 
             if np.any(np.isin(molecule.ati, get_lanthanides())):
                 molecule.uhf = uhf_original
+            if super_heavy_elements:
+                # Reset the atomic numbers to the original values before returning the optimized molecule.
+                molecule.ati = ati_original
             return xtb_log_out
 
     def check_gap(
@@ -188,12 +205,12 @@ class XTB(QMMethod):
                 check=True,
             )
             # get the output of the xtb calculation (of both stdout and stderr)
-            xtb_log_out = xtb_out.stdout.decode("utf8")
-            xtb_log_err = xtb_out.stderr.decode("utf8")
+            xtb_log_out = xtb_out.stdout.decode("utf8", errors="replace")
+            xtb_log_err = xtb_out.stderr.decode("utf8", errors="replace")
             return xtb_log_out, xtb_log_err, 0
         except sp.CalledProcessError as e:
-            xtb_log_out = e.stdout.decode("utf8")
-            xtb_log_err = e.stderr.decode("utf8")
+            xtb_log_out = e.stdout.decode("utf8", errors="replace")
+            xtb_log_err = e.stderr.decode("utf8", errors="replace")
             return xtb_log_out, xtb_log_err, e.returncode