+++ CHANGELOG.md
@@ -16,11 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unit conversion for (currenly unused) vdW radii from the original Fortran project
 - minor print output issues (no new line breaks, more consistent verbosity differentiation, ...)
 - bug in `postprocess_mol` which led to an unassigned return variable in the single-point case
+- bug with all atom lists being initialized with a length of 102 instead of 103
 
 ### Added
 - Support for the novel "g-xTB" method (working title: GP3-xTB)
 - A function which contracts the coordinates after the initial generation.
 - A function which is able to printout the xyz coordinates to the terminal similar to the `.xyz` layout.
+- Elements 87 to 103 are accessible via the element composition. If `xtb` is the engine, the elements will be replaced by their lighter homologues.
 
 ### Breaking Changes
 - Removal of the `dist_threshold` flag and in the `-toml` file.